+++ src/app/pages/reports/developers/developers.component.js
@@ -9,59 +9,66 @@ export const ReportsDevelopersComponent = {
             this.ReportService = ReportService;
             this.networkService = networkService;
             this.utilService = utilService;
-            this.activityRange = {
-                range: 30,
-                startDate: new Date(),
-                endDate: new Date(),
-            };
-            this.activityRange.startDate.setDate(this.activityRange.endDate.getDate() - this.activityRange.range + 1); // offset to account for inclusion of endDate in range
+
+            this.results = [];
+            this.displayed = [];
+            this.clearFilterHs = [];
+            this.restoreStateHs = [];
             this.filename = 'Reports_' + new Date().getTime() + '.csv';
             this.filterText = '';
             this.tableController = {};
+            this.loadProgress = {
+                total: 0,
+                complete: 0,
+            };
+            this.downloadProgress = { complete: 0 };
+            this.pageSize = 50;
         }
 
         $onInit () {
             this.search();
         }
 
-        dateAdjust (obj) {
-            var ret = angular.copy(obj);
-            ret.startDate = this.ReportService.coerceToMidnight(ret.startDate);
-            ret.endDate = this.ReportService.coerceToMidnight(ret.endDate, true);
-            return ret;
-        }
-
-        downloadReady () {
-            if (this.displayed) {
-                return this.displayed.reduce((acc, activity) => activity.action && acc, true);
-            } else {
-                return false;
-            }
+        $onDestroy () {
+            this.isDestroyed = true;
         }
 
         onApplyFilter (filterObj) {
             let f = angular.fromJson(filterObj);
-            this.activityRange.startDate = new Date(Date.parse(f.startDate));
-            this.activityRange.endDate = new Date(Date.parse(f.endDate));
-            this.filterText = f.dataFilter;
-            this.tableController.sortBy(f.tableState.sort.predicate, f.tableState.sort.reverse);
-            this.search();
+            this.doFilter(f)
         }
 
         onClearFilter () {
-            this.activityRange.endDate = new Date();
-            this.activityRange.startDate = this.utilService.addDays(this.activityRange.endDate, (this.activityRange.range * -1) + 1)
-            this.filterText = '';
-            this.tableController.sortBy('date');
-            this.search();
+            let filterData = {};
+            filterData.dataFilter = '';
+            filterData.tableState = this.tableController.tableState();
+            this.clearFilterHs.forEach(handler => handler());
+            this.doFilter(filterData);
+        }
+
+        doFilter (filter) {
+            let that = this;
+            this.filterText = filter.dataFilter;
+            if (filter.tableState.search.predicateObject.date) {
+                this.tableController.search(filter.tableState.search.predicateObject.date, 'date');
+            } else {
+                this.tableController.search({}, 'date');
+            }
+            this.restoreStateHs.forEach(handler => handler(that.tableController.tableState()));
+            this.tableController.sortBy(filter.tableState.sort.predicate, filter.tableState.sort.reverse);
+        }
+
+        registerClearFilter (handler) {
+            this.clearFilterHs.push(handler);
+        }
+
+        registerRestoreState (handler) {
+            this.restoreStateHs.push(handler);
         }
 
         createFilterDataObject () {
             let filterData = {};
-            filterData.startDate = this.ReportService.coerceToMidnight(this.activityRange.startDate);
-            filterData.endDate = this.ReportService.coerceToMidnight(this.activityRange.endDate);
             filterData.dataFilter = this.filterText;
-            filterData.tableState = {};
             filterData.tableState = this.tableController.tableState();
             return filterData;
         }
@@ -77,6 +84,7 @@ export const ReportsDevelopersComponent = {
                     {key: 'developerCode', display: 'Developer Code'},
                     {key: 'name', display: 'Name'},
                     {key: 'website', display: 'Website'},
+                    {key: 'selfDeveloper', display: 'Self-developer'},
                 ];
                 var nestedKeys = [
                     {key: 'status', subkey: 'statusName', display: 'Developer Status'},
@@ -178,7 +186,6 @@ export const ReportsDevelopersComponent = {
                     if (foundEvents) {
                         activity.details.push(translatedEvents);
                     }
-
                 } else if (item.originalData && angular.isArray(item.originalData) && item.newData && !angular.isArray(item.newData)) { // merge
                     activity.action ='Developers ' + item.originalData.map(d => d.name).join(' and ') + ' merged to form ' + item.newData.name;
                     activity.details = [];
@@ -191,41 +198,84 @@ export const ReportsDevelopersComponent = {
                     activity.details = [];
                     activity.csvAction = activity.action.replace(',','","');
                 }
-
                 meta.action = activity.action;
                 meta.details = activity.details;
                 meta.csvDetails = activity.details.join('\n');
             });
         }
 
-        prepare (results) {
-            this.displayed = results.map(item => {
-                item.filterText = item.developerName + '|' + item.developerCode + '|' + item.responsibleUser.fullName
-                item.categoriesFilter = '|' + item.categories.join('|') + '|';
-                item.friendlyActivityDate = new Date(item.date).toISOString().substring(0, 10);
-                item.fullName = item.responsibleUser.fullName;
-                return item;
-            });
+        prepare (item) {
+            item.filterText = item.developerName + '|' + item.developerCode + '|' + item.responsibleUser.fullName;
+            if (item.categories.length > 1 || item.categories[0] !== 'DEVELOPER') {
+                this.$log.info(item.categories);
+            }
+            item.categoriesFilter = '|' + item.categories.join('|') + '|';
+            item.friendlyActivityDate = new Date(item.date).toISOString().substring(0, 10);
+            item.fullName = item.responsibleUser.fullName;
+            return item;
+        }
+
+        canDownload () {
+            return this.displayed
+                .filter(item => !item.action).length <= 1000;
         }
 
         prepareDownload () {
+            let total = this.displayed
+                .filter(item => !item.action).length;
+            let progress = 0;
             this.displayed
                 .filter(item => !item.action)
-                .forEach(item => this.parse(item));
+                .forEach(item => {
+                    this.parse(item).then(() => {
+                        progress += 1;
+                        this.downloadProgress.complete = Math.floor(100 * ((progress + 1) / total));
+                    });
+                });
             //todo, eventually: use the $q.all function as demonstrated in product history eye
+        }
+
+        showLoadingBar () {
+            let tableState = this.tableController.tableState && this.tableController.tableState();
+            return this.ReportService.showLoadingBar(tableState, this.results, this.loadProgress);
         }
 
         search () {
             let that = this;
-            this.networkService.getActivityMetadata('developers', this.dateAdjust(this.activityRange))
+            this.networkService.getActivityMetadata('beta/developers')
                 .then(results => {
-                    that.results = results;
-                    that.prepare(that.results);
+                    that.results = results.activities
+                        .map(item => that.prepare(item));
+                    that.loadProgress.total = (Math.floor(results.resultSetSize / results.pageSize) + (results.resultSetSize % results.pageSize === 0 ? 0 : 1))
+                    let filter = {};
+                    filter.dataFilter = '';
+                    filter.tableState = this.tableController.tableState();
+                    filter.tableState.search = {
+                        predicateObject: {
+                            date: {
+                                after: new Date('2016-04-01').getTime(),
+                                before: this.ReportService.coerceToMidnight(new Date(), true).getTime(),
+                            },
+                        },
+                    };
+                    that.doFilter(filter);
+                    that.addPageToData(1);
                 });
         }
 
-        validDates () {
-            return this.ReportService.validDates(this.activityRange.startDate, this.activityRange.endDate, this.activityRange.range, false);
+        addPageToData (page) {
+            let that = this;
+            if (this.isDestroyed) { return }
+            this.networkService.getActivityMetadata('beta/developers', {pageNum: page, ignoreLoadingBar: true}).then(results => {
+                results.activities.forEach(item => {
+                    that.results.push(that.prepare(item));
+                });
+                that.loadProgress.complete += 1;
+                that.loadProgress.percentage = Math.floor(100 * ((that.loadProgress.complete + 1) / that.loadProgress.total));
+                if (page < that.loadProgress.total) {
+                    that.addPageToData(page + 1);
+                }
+            });
         }
 
         compareTransparencyAttestations (before, after) {
@@ -234,27 +284,26 @@ export const ReportsDevelopersComponent = {
             //on the acbs
             before.forEach(beforeTA => {
                 let afterTA = after.find(ta => ta.acbId === beforeTA.acbId);
-                if (afterTA) {
-                    changes.push(this.compareTransparencyAttestation(beforeTA, afterTA));
+                let change = this.compareTransparencyAttestation(beforeTA, afterTA);
+                if (change) {
+                    changes.push(change);
                 }
             });
             return changes;
         }
 
         compareTransparencyAttestation (before, after) {
-            if (!before.transparencyAttestation && !after.transparencyAttestation) {
-                return '';
-            } else if (!before.transparencyAttestation) {
+            if (!before.transparencyAttestation && after.transparencyAttestation) {
                 //Transparency attestation was added
                 return '<li>Transparency Attestation "' + after.acbName + '" changes<ul><li>Transparency Attestation added: ' + after.transparencyAttestation.transparencyAttestation + '.</li></ul></li>';
-            } else if (!after.transparencyAttestation) {
+            }
+            if (before.transparencyAttestation && !after.transparencyAttestation) {
                 //Transparency attestation was removed - not sure this is possible
                 return '<li>Transparency Attestation "' + after.acbName + '" changes<ul><li>Transparency Attestation removed. Was: ' + before.transparencyAttestation.transparencyAttestation + '.</li></ul></li>';
-            } else if (before.transparencyAttestation.transparencyAttestation !== after.transparencyAttestation.transparencyAttestation) {
+            }
+            if (before.transparencyAttestation && after.transparencyAttestation && before.transparencyAttestation.transparencyAttestation !== after.transparencyAttestation.transparencyAttestation) {
                 //Transparency attestation was changed
                 return '<li>Transparency Attestation "' + after.acbName +'" changes<ul><li>Transparency Attestation changed: ' + after.transparencyAttestation.transparencyAttestation + '. Was: ' + before.transparencyAttestation.transparencyAttestation + '.</li></ul></li>';
-            } else {
-                return '';
             }
         }