+++ src/app/pages/resources/cms-lookup/cms-lookup.module.js
@@ -1,5 +1,0 @@
-(function () {
-    'use strict';
-
-    angular.module('chpl.cms_lookup', ['ngStorage', 'chpl.services']);
-})();