+++ src/app/pages/resources/overview/overview.module.js
@@ -1,6 +1,0 @@
-(function () {
-    'use strict';
-
-    angular
-        .module('chpl.overview', ['chpl.services']);
-})();