+++ src/app/services/network.service.js
@@ -290,14 +290,14 @@ export class NetworkService {
         case 'complaintListings':
             return this.apiGET('/collections/certified_products?fields=id,acb,chplProductNumber,developer,product');
         case 'correctiveAction':
-            return this.apiGET('/collections/certified_products?fields=id,edition,developer,product,version,chplProductNumber,certificationStatus,acb,surveillanceCount,openNonconformityCount,closedNonconformityCount');
+            return this.apiGET('/collections/certified_products?fields=id,edition,curesUpdate,developer,product,version,chplProductNumber,certificationStatus,acb,surveillanceCount,openNonconformityCount,closedNonconformityCount');
         case 'decertifiedProducts':
         case 'inactiveCertificates':
-            return this.apiGET('/collections/certified_products?fields=id,edition,developer,product,version,chplProductNumber,acb,decertificationDate,certificationStatus,numMeaningfulUse,numMeaningfulUseDate');
+            return this.apiGET('/collections/certified_products?fields=id,edition,curesUpdate,developer,product,version,chplProductNumber,acb,decertificationDate,certificationStatus,numMeaningfulUse,numMeaningfulUseDate');
         case 'sed':
             return this.apiGET('/collections/certified_products?fields=id,edition,developer,product,version,chplProductNumber,acb,certificationStatus,criteriaMet');
         case 'surveillanceManagement':
-            return this.apiGET('/collections/certified_products?fields=id,edition,developer,product,version,chplProductNumber,certificationStatus,acb,openSurveillanceCount,closedSurveillanceCount,openNonconformityCount,closedNonconformityCount,surveillanceDates');
+            return this.apiGET('/collections/certified_products?fields=id,edition,curesUpdate,developer,product,version,chplProductNumber,certificationStatus,acb,openSurveillanceCount,closedSurveillanceCount,openNonconformityCount,closedNonconformityCount,surveillanceDates');
         case 'transparencyAttestations':
             return this.apiGET('/collections/developers');
             //no default